jobs/generate_exam/pyproject.toml
[project]
name = "generate-exam"
version = "0.1.0"
description = "Job to generate exams based on lecture materials"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.38.5",
    "botocore>=1.38.5",
    "httpx==0.27.2",
    "openai>=1.76.2",
    "psycopg2>=2.9.10",
    "pymupdf>=1.25.5",
    "pyyaml>=6.0.2",
    "pydantic>=2.0.0",
]
